[connectivity] Move as much logic as possible to the MethodChannel implementation.
connectivity_platform_interface:

* Bring ConnectivityResult and LocationAuthorizationStatus enums from
core package.
* Use the above Enums as return values for ConnectivityPlatformInterface
methods.
* Modify the MethodChannel implementation so it returns the right types.
* Bring utility methods, asserts and other logic that is only needed on
the MethodChannel implementation from the core package, so it's simpler.
* Bring MethodChannel unit tests from core package.

connectivity (core pkg):

* Reexport ConnectivityResult and LocationAuthorizationStatus enums so
plugin users can see them.
* Remove MethodChannel unit tests (moved to platform_interface).
