+++ packages/connectivity/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.4.7
+
+* Migrate the plugin to use the ConnectivityMethodChannel implementation from the connectivity_platform_interface package.
+
 ## 0.4.6+2
 
 * Migrate deprecated BinaryMessages to ServicesBinding.instance.defaultBinaryMessenger.
